# docs/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: HSE Platform API
  description: "\nHSE platform API — Phase 0 Foundation (spec `docs/specs/0-foundation.md` v1.0).\n\n\
    * Auth: `POST /api/v1/auth/login` sets the httpOnly SameSite=Lax cookie `hse_session` (JWT) and\n\
    \  returns `{access_token, user}`. Send the cookie or `Authorization: Bearer <token>`.\n* Errors:\
    \ `{\"detail\": {\"code\", \"message\", \"message_ar\", \"errors\"}}`; 401 unauthenticated,\n  403\
    \ forbidden, 404 not found / out of scope, 409 conflict / invalid transition, 422 validation.\n* Lists:\
    \ `?page=1&page_size=50` → `{items, total, page, page_size}`.\n* Timestamps are UTC; display and day\
    \ boundaries use the project timezone (Asia/Riyadh).\n"
  version: 0.1.0
paths:
...
  /api/v1/contractors:
    get:
      tags:
      - contractors
      summary: List contractors visible to the caller
      description: '`q` matches short code, CR and EN/AR names with Arabic normalisation (rule 45). Contact
        fields are omitted without capability 9.'
      operationId: list_contractors
      security:
      - sessionCookie: []
      - bearerAuth: []
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: array
            items:
              $ref: '#/components/schemas/ContractorStatus'
          - type: 'null'
          title: Status
      - name: category
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/ContractorCategory'
          - type: 'null'
          title: Category
      - name: project_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          description: Only contractors engaged on this project.
          title: Project Id
        description: Only contractors engaged on this project.
      - name: cr_expiring_within_days
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
            maximum: 365
            minimum: 0
          - type: 'null'
          description: CR expiry within N days (incl. expired).
          title: Cr Expiring Within Days
        description: CR expiry within N days (incl. expired).
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 100
          - type: 'null'
          title: Q
      - name: sort
        in: query
        required: false
        schema:
          enum:
          - short_code
          - -short_code
          - name
          - -name
          - cr_expiry_date
          type: string
          default: short_code
          title: Sort
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          description: 1-based page number.
          default: 1
          title: Page
        description: 1-based page number.
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          description: Items per page.
          default: 50
          title: Page Size
        description: Items per page.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractorPage'
        '401':
          description: Not authenticated, session expired, or login rejected.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Missing capability, read-only role, contractor suspended, or privacy notice not
            acknowledged (PRIVACY_ACK_REQUIRED).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error (field errors in detail.errors).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
      - contractors
      summary: Create a contractor (draft)
      operationId: create_contractor
      security:
      - sessionCookie: []
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContractorCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractorRead'
        '401':
          description: Not authenticated, session expired, or login rejected.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Missing capability, read-only role, contractor suspended, or privacy notice not
            acknowledged (PRIVACY_ACK_REQUIRED).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Invalid state transition, closed project, duplicate value or business-rule conflict.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error (field errors in detail.errors).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/contractors/{contractor_id}:
    get:
      tags:
      - contractors
      summary: Get a contractor
      operationId: get_contractor
      security:
      - sessionCookie: []
      - bearerAuth: []
      parameters:
      - name: contractor_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Contractor Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractorRead'
        '401':
          description: Not authenticated, session expired, or login rejected.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Missing capability, read-only role, contractor suspended, or privacy notice not
            acknowledged (PRIVACY_ACK_REQUIRED).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found or outside the caller's project/contractor scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error (field errors in detail.errors).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags:
      - contractors
      summary: Update a contractor
      operationId: update_contractor
      security:
      - sessionCookie: []
      - bearerAuth: []
      parameters:
      - name: contractor_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Contractor Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContractorUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractorRead'
        '401':
          description: Not authenticated, session expired, or login rejected.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Missing capability, read-only role, contractor suspended, or privacy notice not
            acknowledged (PRIVACY_ACK_REQUIRED).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found or outside the caller's project/contractor scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Invalid state transition, closed project, duplicate value or business-rule conflict.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error (field errors in detail.errors).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/contractors/{contractor_id}/transitions:
    post:
      tags:
      - contractors
      summary: Change contractor status (submit, approve, return, suspend, reinstate, demobilise, blacklist,
        lift blacklist)
      description: Blacklisting deactivates the contractor's users, ends their sessions and role assignments,
        and flags descendant engagements (rule 27).
      operationId: transition_contractor
      security:
      - sessionCookie: []
      - bearerAuth: []
      parameters:
      - name: contractor_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Contractor Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContractorTransitionRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractorRead'
        '401':
          description: Not authenticated, session expired, or login rejected.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Missing capability, read-only role, contractor suspended, or privacy notice not
            acknowledged (PRIVACY_ACK_REQUIRED).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found or outside the caller's project/contractor scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Invalid state transition, closed project, duplicate value or business-rule conflict.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error (field errors in detail.errors).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/projects/{project_id}/engagements:
    get:
      tags:
      - contractors
      summary: List contractor engagements on a project (scoped to the contractor tree for contractor
        roles)
      operationId: list_engagements
      security:
      - sessionCookie: []
      - bearerAuth: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      - name: tier
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
            maximum: 3
            minimum: 1
          - type: 'null'
          title: Tier
      - name: contractor_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Contractor Id
      - name: parent_engagement_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Parent Engagement Id
      - name: site_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Site Id
      - name: parent_blacklisted
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Parent Blacklisted
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          description: 1-based page number.
          default: 1
          title: Page
        description: 1-based page number.
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          description: Items per page.
          default: 50
          title: Page Size
        description: Items per page.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EngagementPage'
        '401':
          description: Not authenticated, session expired, or login rejected.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Missing capability, read-only role, contractor suspended, or privacy notice not
            acknowledged (PRIVACY_ACK_REQUIRED).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found or outside the caller's project/contractor scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error (field errors in detail.errors).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
      - contractors
      summary: Engage an approved contractor on a project
      description: 409 CONTRACTOR_NOT_APPROVED unless the contractor is approved; 422 when the parent
        tier is not tier-1 or is on another project.
      operationId: create_engagement
      security:
      - sessionCookie: []
      - bearerAuth: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EngagementCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EngagementRead'
        '401':
          description: Not authenticated, session expired, or login rejected.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Missing capability, read-only role, contractor suspended, or privacy notice not
            acknowledged (PRIVACY_ACK_REQUIRED).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found or outside the caller's project/contractor scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Invalid state transition, closed project, duplicate value or business-rule conflict.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error (field errors in detail.errors).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/engagements/{engagement_id}:
    get:
      tags:
      - contractors
      summary: Get an engagement
      operationId: get_engagement
      security:
      - sessionCookie: []
      - bearerAuth: []
      parameters:
      - name: engagement_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Engagement Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EngagementRead'
        '401':
          description: Not authenticated, session expired, or login rejected.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Missing capability, read-only role, contractor suspended, or privacy notice not
            acknowledged (PRIVACY_ACK_REQUIRED).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found or outside the caller's project/contractor scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error (field errors in detail.errors).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags:
      - contractors
      summary: Update an engagement (scope, sites, dates)
      operationId: update_engagement
      security:
      - sessionCookie: []
      - bearerAuth: []
      parameters:
      - name: engagement_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Engagement Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EngagementUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EngagementRead'
        '401':
          description: Not authenticated, session expired, or login rejected.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Missing capability, read-only role, contractor suspended, or privacy notice not
            acknowledged (PRIVACY_ACK_REQUIRED).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found or outside the caller's project/contractor scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Invalid state transition, closed project, duplicate value or business-rule conflict.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error (field errors in detail.errors).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/engagements/{engagement_id}/clear-parent-blacklisted:
    post:
      tags:
      - contractors
      summary: 'HSE Manager: mark a ''parent blacklisted'' flag as reviewed'
      operationId: clear_parent_blacklisted
      security:
      - sessionCookie: []
      - bearerAuth: []
      parameters:
      - name: engagement_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Engagement Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EngagementRead'
        '401':
          description: Not authenticated, session expired, or login rejected.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Missing capability, read-only role, contractor suspended, or privacy notice not
            acknowledged (PRIVACY_ACK_REQUIRED).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found or outside the caller's project/contractor scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Invalid state transition, closed project, duplicate value or business-rule conflict.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error (field errors in detail.errors).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ContractorCategory:
      type: string
      enum:
      - civil
      - mep
      - steel
      - airfield
      - scaffolding
      - lifting
      - facade
      - specialist
      - consultant
      - other
      title: ContractorCategory
    ContractorCreate:
      properties:
        legal_name_en:
          type: string
          maxLength: 200
          minLength: 1
          title: Legal Name En
        legal_name_ar:
          type: string
          maxLength: 200
          minLength: 1
          title: Legal Name Ar
        short_code:
          type: string
          pattern: ^[A-Z0-9]{2,10}$
          title: Short Code
          description: Unique, e.g. RAWABI.
        cr_number:
          type: string
          pattern: ^[12457]\d{9}$
          title: Cr Number
          description: 10-digit Commercial Registration; legacy prefix 1/2/4/5 or unified 7. Unique.
        cr_expiry_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Cr Expiry Date
        vat_number:
          anyOf:
          - type: string
            pattern: ^3\d{13}3$
          - type: 'null'
          title: Vat Number
        contractor_category:
          $ref: '#/components/schemas/ContractorCategory'
        primary_contact_name:
          type: string
          maxLength: 120
          minLength: 1
          title: Primary Contact Name
        primary_contact_mobile:
          type: string
          maxLength: 16
          title: Primary Contact Mobile
          description: E.164; KSA +9665XXXXXXXX.
        primary_contact_email:
          type: string
          format: email
          title: Primary Contact Email
      additionalProperties: false
      type: object
      required:
      - legal_name_en
      - legal_name_ar
      - short_code
      - cr_number
      - contractor_category
      - primary_contact_name
      - primary_contact_mobile
      - primary_contact_email
      title: ContractorCreate
      description: Creates a contractor in status draft (capability 5).
    ContractorPage:
      properties:
        items:
          items:
            $ref: '#/components/schemas/ContractorRead'
          type: array
          title: Items
        total:
          type: integer
          title: Total
          description: Total number of matching items across all pages.
        page:
          type: integer
          title: Page
          description: 1-based page number.
        page_size:
          type: integer
          title: Page Size
          description: Items per page (max 200).
      type: object
      required:
      - items
      - total
      - page
      - page_size
      title: ContractorPage
    ContractorRead:
      properties:
        created_at:
          type: string
          format: date-time
          title: Created At
          description: UTC timestamp.
        updated_at:
          type: string
          format: date-time
          title: Updated At
          description: UTC timestamp.
        id:
          type: string
          format: uuid
          title: Id
        legal_name_en:
          type: string
          title: Legal Name En
        legal_name_ar:
          type: string
          title: Legal Name Ar
        short_code:
          type: string
          title: Short Code
        cr_number:
          type: string
          title: Cr Number
        cr_expiry_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Cr Expiry Date
        vat_number:
          anyOf:
          - type: string
          - type: 'null'
          title: Vat Number
        contractor_category:
          $ref: '#/components/schemas/ContractorCategory'
        status:
          $ref: '#/components/schemas/ContractorStatus'
        status_reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Status Reason
        primary_contact_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Primary Contact Name
          description: Personal. Omitted when the caller lacks capability 9 (view contractor contact details).
        primary_contact_mobile:
          anyOf:
          - type: string
          - type: 'null'
          title: Primary Contact Mobile
          description: Personal. Omitted when the caller lacks capability 9 (view contractor contact details).
        primary_contact_email:
          anyOf:
          - type: string
          - type: 'null'
          title: Primary Contact Email
          description: Personal. Omitted when the caller lacks capability 9 (view contractor contact details).
      type: object
      required:
      - created_at
      - updated_at
      - id
      - legal_name_en
      - legal_name_ar
      - short_code
      - cr_number
      - cr_expiry_date
      - vat_number
      - contractor_category
      - status
      - status_reason
      title: ContractorRead
    ContractorStatus:
      type: string
      enum:
      - draft
      - pending_approval
      - approved
      - suspended
      - demobilised
      - blacklisted
      title: ContractorStatus
      description: Contractor master-record lifecycle (§4.2).
...
    ContractorUpdate:
      properties:
        legal_name_en:
          anyOf:
          - type: string
            maxLength: 200
            minLength: 1
          - type: 'null'
          title: Legal Name En
        legal_name_ar:
          anyOf:
          - type: string
            maxLength: 200
            minLength: 1
          - type: 'null'
          title: Legal Name Ar
        short_code:
          anyOf:
          - type: string
            pattern: ^[A-Z0-9]{2,10}$
          - type: 'null'
          title: Short Code
        cr_number:
          anyOf:
          - type: string
            pattern: ^[12457]\d{9}$
          - type: 'null'
          title: Cr Number
        cr_expiry_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Cr Expiry Date
        vat_number:
          anyOf:
          - type: string
            pattern: ^3\d{13}3$
          - type: 'null'
          title: Vat Number
        contractor_category:
          anyOf:
          - $ref: '#/components/schemas/ContractorCategory'
          - type: 'null'
        primary_contact_name:
          anyOf:
          - type: string
            maxLength: 120
            minLength: 1
          - type: 'null'
          title: Primary Contact Name
        primary_contact_mobile:
          anyOf:
          - type: string
            maxLength: 16
          - type: 'null'
          title: Primary Contact Mobile
        primary_contact_email:
          anyOf:
          - type: string
            format: email
          - type: 'null'
          title: Primary Contact Email
      additionalProperties: false
      type: object
      title: ContractorUpdate
...
    EngagementPage:
      properties:
        items:
          items:
            $ref: '#/components/schemas/EngagementRead'
          type: array
          title: Items
        total:
          type: integer
          title: Total
          description: Total number of matching items across all pages.
        page:
          type: integer
          title: Page
          description: 1-based page number.
        page_size:
          type: integer
          title: Page Size
          description: Items per page (max 200).
      type: object
      required:
      - items
      - total
      - page
      - page_size
      title: EngagementPage
...
    ErrorResponse:
      properties:
        detail:
          $ref: '#/components/schemas/ErrorDetail'
      type: object
      required:
      - detail
      title: ErrorResponse
      description: Envelope for every 4xx/5xx response.
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      description: httpOnly SameSite=Lax cookie set by POST /auth/login.
      in: cookie
      name: hse_session
    bearerAuth:
      type: http
      description: 'Same JWT as the cookie, as `Authorization: Bearer <token>`.'
      scheme: bearer